test: add coverage for make_tool_button?
