+++ tests/conftest.py
@@ -4,8 +4,20 @@ import sys
 import os
 from pathlib import Path
 
+import pytest
+from PySide6.QtWidgets import QApplication
+
 ROOT = Path(__file__).resolve().parents[1]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
 os.environ["QT_QPA_PLATFORM"] = "offscreen"
+
+
+@pytest.fixture(scope="module")
+def _app():
+    """Provide a ``QApplication`` instance for UI tests."""
+    qapp = QApplication.instance()
+    if qapp is None:
+        qapp = QApplication([])
+    return qapp
